Updated openstack/openstack Project: openstack/horizon  37e064a33f980a0e59e756ec3a067bcdd85ae141
Not able to delete a pseudo-folder via horizon

Through horizon the user not able to delete a
pseudo-folder. it giving a error message like
'you are not allowed to delete object'

problem-1:
'subfolders' datatype is not added in the
allowed datatypes tuple, so the delete operation
is not allowed for pseudo-folder. now the 'subfolder'
is added in the tuple.

problem-2:
the object name is starting with container name and
not trailing with '/'. now the object name is modified
before sending to swift client.(removing the container
name and trailing with '/'.

Change-Id: Ifced62b24a9b6e4800b70dae2912ff16d809743d
Closes-Bug: #1317016
Co-Authored-By: Mizielski Robert <robert.mizielski@cloudwatt.com>