Updated openstack/openstack Project: openstack/horizon  33367b12c7055a109c0ab4655186d04f0a00a353
Add encryption indicator for volumes

This patch adds support to horizon for a volume encryption
indicator, to display to the user when a volume has an encrypted
volume type. It uses a change made in cinder
(https://review.openstack.org/#/c/72483/) to determine whether
a volume is encrypted or not.

The patch adds an "Encrypted?" column to the project volumes
table, which has the value "Yes" if the volume is encrypted,
and "No" otherwise.  This column is also included in the
admin volumes table.

Implements: blueprint integration-with-cinder-volume-encryption

Change-Id: I2913b96a704d1183606f9fdd970b14a02956e2bb