Updated openstack/openstack Project: openstack/nova  6cbc9ee9cf2cb82707a8d89c989127b0b4f0f81c
Fix object change detection

The current object change detection actually just hashed the method
argspecs of the remotable decorators instead of the wrapped methods
themselves. This patch fixes that by stashing a pointer to the
original function, which is used for the hash and will be useful
for other inspection-related tasks as well.

This also adds a utility to help when regenerating the hashes,
similar to what we do for API samples. Running the unit tests like
this will give you an object_hashes.txt output file:

  GENERATE_HASHES=y tox TestObjectVersions

Change-Id: Icfe4d6c08f42296c756d1f9ab7c88b52e3209d84