Updated openstack/openstack Project: openstack/nova  ab3e18d3f3fe5069dfc5e3d36d17a9aef7f24453
Network: add instance to the debug statement

Allocating and deallocating a network from an instance needs
more info to debug especially when more than one instance
has the same name.

The same change is also made regarding exceptions. That is,
the instance UUID will be used instead of the name.

Change-Id: I624fe77c71f6cb7ca453c863b59f0452c8e6b383