Updated openstack/openstack Project: openstack/cinder  2eb2294c69ca3ddc96d8b8955ad605ab162f2c8b
Made provision for providing optional arguments

The 'quota_committed' attribute of 'RequestContext' object is
a transient property, so it will not be saved in the taskflow
persistent storage. The updated value of 'quota_committed'
attribute will not be available while resuming/reverting the
flow, if cinder api-service is down/stopped after committing
the quota.

Since this 'quota_committed' attribute is not used anywhere
in cinder project other than in create-volume taskflow api, so
removed 'quota_committed' from RequestContext and made
provision to pass it as an optional argument which will be
passed to api-flow via create_what dictionary, in order to
make it persistent and use it as and when needed.

Change-Id: I719067a8e1824add45fb525c106642a9c0e6bb46