Updated openstack/openstack Project: openstack/heat  a5b78a2e0a17198bbf86840bd974a4dff53bb38b
Assume compute service_type for nova client

This change removes the ability to specify a non-compute service
type when requesting a nova client. The only non-compute service
type actually requested was the fallback path for cinder. However
nova volumes was completely removed from nova in grizzly, so we
do not need to support this any more.

Change-Id: I9e15f31979d44f018db69ebc8109f707b0f51671