Updated openstack/openstack Project: openstack-dev/devstack  3723814bf27fb4d78c6c3ad80d77882f75ad07c4
convert devstack from awk to outfilter

The complexity of the awk script was pretty high, and it would
be good to do it in summaries as well, which starts to get a bit
squirelly. Instead bring over the outfilter.py from grenade and
use it for the timestamping.

Any additional overhead from python should be offset from not
shelling out to date on every line of output.

Change-Id: Ic2b86ddba3e7f6520a0fd35599b01143936c6deb