Updated openstack/openstack Project: openstack-dev/devstack  56a564021d317004a6e75167cc5114ea49355356
double the screen sleep timeout

HP Cloud 1.1 is quite slow per core. So much so that grenade is
often failing to start services because bash in a screen session
isn't ready in 1.5s. Double this to 3s to try to decrease failures
in the gate.

We should instead do this via some readiness check to decrease
races.

Change-Id: I2b066a687916742e966190f00b5b06a795d4a014