Updated openstack/openstack Project: openstack/nova  96b8e579bda0412abcd6e4012e79f1d3dc075142
Skip migrations test_walk_versions instead of pass

When there is no db engine configured for the migration tests,
the test_walk_versions test was passing, rather than skipping,
because it contained a for-loop and no check for the length of the
sequence being iterated over.
This patch checks the length of the configured engines and
properly skips the test when there are no engines.

Change-Id: I4078655f01638c22cf029f088018b8be93d6a01a
Co-Authored-By: Devananda van der Veen<devananda.vdv@gmail.com>
Closes-bug: #1327397