Updated openstack/openstack Project: openstack/tempest  f064d3e37e4861fd848a70b7128f415c4a3322cc
Clarify API extension description of Nova API

On Nova API, we can see two names for each API extension. The one is
"Name", the other is "Alias". The "Name" is represented with camelcase
and the "Alias" is done with characters and hyphens:

 $ nova list-extensions
 +--------------+-----------------------+------------------+---------+
 | Name         | Summary               | Alias            | Updated |
 +--------------+-----------------------+------------------+---------+
 | AdminActions | Enable admin-only ... | os-admin-actions | ...     |
 | Agents       | Agents support.       | os-agents        | ...     |
 | ConsoleOutput| Console log output... | os-console-output| ...     |
 | Consoles     | Interactive Console.. | os-consoles      | ...     |

In Tempest, we can select API extensions what we test and it should be
based on "Alias", because on Nova side we have been a lot stricter about
format/uniqueness for the alias which is seen as the canonical name.
This patch clarifies it.

Change-Id: I3e753c81182af83e01ad3a1f2eb9782c8abb6f50