Updated openstack/openstack Project: openstack/requirements  d120190f00545a8768a10d803f94ea43aebd09ba
Fix whitespace that occurs after comma

Use a custom tuple of separators for json.dumps that avoids
having the comma+space, which looks bad if this is ever committed.
Replace it by just having ": " and "," (no space after) instead.

An example of this so that it can be more clear:

http://paste.openstack.org/show/84238/

Note the above paste where the old version has spaces after
commas where as the newer version does not have these spaces.

Change-Id: Ifd04d5e1b26ebdf9db374761f31baaf1e33d87cd