Updated openstack/openstack Project: openstack/horizon  3a9d4658306158d9f3b74ee708faef28d48a74b8
Sort security group rules by protocol and port

When adding new rules to a security group, it is useful to have the
rules sorted consistently across page reloads. This patch ensures
that the rules are sorted by protocol first, then by port. It is still
possible to manually sort the table using the client side capabilities,
but by default the table is now sorted by protocol and port, server
side.

Change-Id: I36d29461a8437f5689425706a7faae3399c6bf23
Closes-bug: #1326132