Updated openstack/openstack Project: openstack/tempest  ffff29e243440f2b2b9996ebd441ea5468940148
Add unittests for remote client

Remote client is used in servers tests to provide some common operations into a
remote server. It's backended by the ssh class, which already has tests, so all
of its calls are mocked and due to that there are some 'useless' tests that are
pretty much checking response's format and checking if the ssh call didn't
change to avoid inconsistencies in the future.

Partially implements bp unit-tests

Change-Id: I5dabc170a557c81bb9b0605bf101719585d7c53b