Updated openstack/openstack Project: openstack/tempest  d1cbdc107ab70828fdebc4076229fd61e047cba0
Unbreak baremetal API tests that were not updated

A recent merge (88233) changed the baremetal test base to return (resp, body)
tuples from its create_* methods instead of the created objects.  This updates
the test cases that were missed in this patch and unbreaks all Ironic CI jobs
that were overlooked in that review.

Change-Id: I691e36d0678a15737614b7867a4c6778e714f299
Closes-bug: #1330834