Updated openstack/openstack Project: openstack/glance  c6dd4917f29ac7f33711ad8320c2e5b247bd300c
Move `location`'s domain code out of glance.store

This patch moves the location domain's specific code out of
glance.store. Since this code is Glance specific, it's necessary to move
it out of the store package so that we can then remove it all together
and use the `glance.store` library.

Partially-Implements blueprint: create-store-package

Change-Id: I7dc3d5d211054cabc2adc2b9b689cda11bce2190