Updated openstack/openstack Project: openstack/ceilometer  70fa4874ad279abd6d16abde9df697ddc4e42568
Respect $TMPDIR environment variable to run tests

On Fedora 20, /tmp now uses the tmpfs file system which means that all
files are stored in memory. Running Ceilometer tests write 1.2 GB in
/tmp.

This patch allows to set the TMPDIR environment variable to run tests in
a different temporary directoy. So I can use a directory on disk to not
use all my memory.

Change-Id: Ib87a988a4acb3cb568d3f3dda641f5a824f986da