Updated openstack/openstack Project: openstack/sahara  77fb369761dc2a2db2c49df861046eb7a79cc7ae
Handle remote driver not loaded situation

An error in our code could lead to remote being used by sahara-api.
Since remote is not available there, an exception will be thrown like
the one in bug #1325989. We should give user a hint that switching to
sahara-all might fix the issue.

Related-Bug: #1325989
Change-Id: I79a00c225b5fe8f98f413b9d88d223076f9b94c3