Updated openstack/openstack Project: openstack/heat  3b72cdf59b6401174b34d03d05c3bc3a70d5d739
Update: Keep a copy of the old template for rollback

To implement the blueprint update-failure-recovery we will need to modify
the current template as we change the stack. We'll need an unmodified copy
to roll back to in the event of failure, so base the previous stack on a
copy.

Change-Id: Ic06acd3c9a89312d1ad8922e26ab02bee4bafeb0