Updated openstack/openstack Project: openstack/nova  8320559ea511eb54467b0e35237b309813b3c142
Add bulk create/destroy functionality to FloatingIP

This adds bulk create and destroy functionality to the FloatingIP
object, which in turn requires that the DB API return the list of
IPs that it created.

This also fixes a bug in the load routine for FloatingIP, which
would previously try to load a FixedIP object with id=None,
which doesn't make sense.

Related to blueprint compute-manager-objects-juno

Change-Id: Ie3c7a5e672fcdb3e0df3d0e46b905b237dcfa182