Updated openstack/openstack Project: openstack/tempest  5c42268d5053d6e1401d20d6cd2e65410f8f809c
Sync versionutils from oslo-incubator

This is needed for comparing client versions from the command line to
required versions for CLI tests.

Tempest already has the immediate oslo dependencies of gettextutils and
log and those are up-to-date enough for what's needed in versionutils,
so this is a --nodeps sync.

Part of blueprint minversion-check-for-cli-tests

Change-Id: I19784499a9bb1135a9c8dc8b26fc56c53489ed22