Updated openstack/openstack Project: openstack/tempest  3dd988f2fa3bad20043281aa4f70bd21e46495e5
Don't include boto tests as part of smoke tests

Tests with the smoke attribute are used all over when we want to run a
subset of tempest. The smoke job is supposed to be a subset that gives
us the most bang for the buck in functional test validation. The ec2 api
falls outside of that being a third party api its always going to be
extra.

Change-Id: I474ada06064d2b236f52e623f7bc19326e711977