Updated openstack/openstack Project: openstack/tempest  787852ffcad8b99033f54878e5e883f0dff99bf7
Use known flavor ID as marker

The flavors marker test races with the flavors admin test because we
can list the flavors, pick the first flavor ID as a marker, then the
flavors admin test deletes that flavor and finally our attempt to list
flavors with that marker fails.

Avoid this by using a known flavor ID as our marker.

Change-Id: I56d043b58a403fc28a13b6f45540831ae8e0331d
Closes-Bug: #1330830