Updated openstack/openstack Project: openstack/nova  63db47035d6413f15b42301c201cecf465521e25
Update websocketproxy to work with websockify 0.6

Websockify version 0.6 bring with it several bugfixes
and new features that affect Nova (including a fix for
novncproxy zombies hanging around and better support
for the python logging framework).  However, it also
broke backwards compatibility due to a refactor which
brought it inline with other Python socket server libraries.

This patch updates the websockify code to function with
websockify version 0.6 as well as websockify version 0.5.x.

Change-Id: I7858dbb49bee8de3fe725beb6793eac3ecba71b6