Updated openstack/openstack Project: openstack/nova  922a5a652b9d1dc6b6d9bdfb390cd2526af2b4cb
Rename VolumeMapping to EC2VolumeMapping

During the code review for adding the InstanceMapping object, a
request was made by a couple of people to rename it to
EC2InstanceObject.  I did that so now I think we should rename the one
that already merged to also have the EC2 prefix.

Change-Id: I0b0847b6d4e229c973e268045709fa90726baed9