Updated openstack/openstack Project: openstack/requirements  a1f1016ca6a8932a35f77b611d230df8112c022a
Add nodeenv for horizon

nodeenv is a tool to create isolated node.js environments. It is already
used in the tuskar-ui project for testing, and would be introduced in
Horizon in patch https://review.openstack.org/#/c/97237/

It is Python 3 compatible and released under the BSD license.

Change-Id: I17d5e9ac36a7982b21e0fdf70d52f6be5069e850
Implements: blueprint node-toolchain